Extract UM982 into custom UserControl
